requirements.txt
asgiref==3.5.0
boto3==1.20.53
botocore==1.23.53
certifi==2021.10.8
Django==4.0.1
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
